gcp_api/requirements.txt
pip>=22.3
google.cloud.storage
horey.configuration_policy>=1.0.0
horey.h_logger>=1.0.0
horey.common_utils>=1.0.0
horey.network>=1.0.0
